refactor: separate SingleCardLayout from Cards
